requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
google-generativeai==0.8.0
anthropic==0.34.0
httpx==0.27.0
